Removed copilot version specification. Moved evaluation tool to engsys.
