+++ tools/azsdk-cli/Azure.Sdk.Tools.Cli.Tests/Helpers/EngSys/PipelineFixSurvivalJudgeTests.cs
@@ -1,14 +1,14 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 using Azure.Sdk.Tools.Cli.CopilotAgents;
-using Azure.Sdk.Tools.Cli.Helpers.Pipeline;
+using Azure.Sdk.Tools.Cli.Helpers.EngSys;
 using Azure.Sdk.Tools.Cli.Models.Pipeline;
 using Azure.Sdk.Tools.Cli.Services;
 using Azure.Sdk.Tools.Cli.Tests.TestHelpers;
 using Moq;
 using Octokit;
 
-namespace Azure.Sdk.Tools.Cli.Tests.Helpers.Pipeline;
+namespace Azure.Sdk.Tools.Cli.Tests.Helpers.EngSys;
 
 /// <summary>
 /// Seam: <see cref="PipelineFixSurvivalJudge.EvaluateAsync"/> turns a set of Copilot commit SHAs and the
