+++ tools/azsdk-cli/Azure.Sdk.Tools.Cli/Helpers/EngSys/PipelineFixEvaluatorHelper.cs
@@ -6,7 +6,7 @@ using Azure.Sdk.Tools.Cli.Models.Pipeline;
 using Azure.Sdk.Tools.Cli.Services;
 using Octokit;
 
-namespace Azure.Sdk.Tools.Cli.Helpers.Pipeline;
+namespace Azure.Sdk.Tools.Cli.Helpers.EngSys;
 
 /// <summary>
 /// Measures how often Copilot fixes a failing pipeline in a repository over a window of time.
